requirements.txt
flask-babel
flask>=1.1
google-cloud-datastore
opencensus-ext-azure
opencensus-ext-flask

# "testing" packages
pytest
requests
